refactored dash
