+++ src/components/Dashboard/DashboardContainer.tsx
@@ -1,0 +1,14 @@
+import { Container } from "react-bootstrap";
+import Greetings from "./Greetings";
+import PollsList from "./PollsList";
+
+const DashboardContainer = (): JSX.Element => {
+  return (
+    <Container className="outer-container" fluid>
+      <Greetings />
+      <PollsList />
+    </Container>
+  );
+};
+
+export default DashboardContainer;
