+++ src/components/Dashboard/Greetings.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useSelector } from "react-redux";
 import { Button, Row } from "react-bootstrap";
-import { RootState } from "../store/store";
+import { RootState } from "../../store/store";
 
 const Greetings = (): JSX.Element => {
   const displayName = useSelector(
